rename podv2 to v2/pod
