+++ internal/k8s/v2/pod/pod.go
@@ -14,8 +14,8 @@
 // limitations under the License.
 //
 
-// Package podv2 provides kubernetes pod information and preriodically update
-package podv2
+// Package pod provides kubernetes pod information and preriodically update
+package pod
 
 import (
 	"context"
